backend/requirements.txt
fastapi
uvicorn
chromadb
sentence-transformers
python-jose
ollama
